requirements.txt
tokenizers
transformers
pymongo
langchain-community
langchain-core
langchain-ollama
langchain-openai
python-dotenv
pandas
regex
spacy
tqdm
en_core_web_sm@https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz
